apps/agent/requirements.txt
faiss-cpu
google-genai